+++ client/src/components/Ingredients/FormulaItem/style.ts
